Added support for fa-stack-corner-* icons
Added some css that allows the user to stack icons with the class fa-stack-corner-*,
where * can be nw, ne, sw or se. This would mainly be useful for adding a plus or minus sign
to any existing icon. I also added some text shadow to help the corner icon be a little more legible.