Fix URL in README to use with jekyll
The current URL does not work. Also enforced the link destination with a markdown link because the old one was including the colon.

I also noticed that I had to run `npm install less` first. Not sure if that is just my broken local JS environment or the mere fact that I have no idea of all that stuff. ;) Maybe somebody with an actual idea of the JS ecosystem can have a look and decide if this is missing in the docs.