Add Joomla3 support to FontAwesome.
This PR provides support for mapping Joomla3's icon classes to Font Awesome fonts. Where possible FA icons which are visually identical or similar to Joomla's icomoon icons are used, with sematically similar icons used as an alternative. Where there are no icons which are visually or sematically similar the icon is left blank.

The CSS provided reuses much of the code from FA 3.2 used to remove the bootstrap icon classes and to define new icon classes.

The Joomla icons can be implemented either by including a new font-awesome-joomla.css file or by removing the comment from a line in the top-level LESS or SCSS file and recompiling.

Additionally, new top-level LESS and SCSS files are provided for generating font-awesome-joomla.css, and documentation has been added to whats-new and get-started wiki pages.