Resolve typings for unit testing
This allows unit tests to be written on the library. Previously a "ts-ignore" statement would be required for this as the TypeScript Compiler would claim that "definitions" did not exist on "Library", despite the fact that when logging the full "Library" it most certianly did. I've made the property readonly since it should be immutable (as the "add" method should be used to add new definitions) and the index signature added to IconDefinitions is to satisfy the TypeScript compiler when the compiler option of "strict" (or "NoImplicitAny") is enabled.

Signed-off-by: Jeroen Claassens <jeroen.claassens@live.nl>